fix: 관리자 페이지 totalCount long 계약 적용
- 실패 수집 조회와 외부 API 로그 조회 응답의 totalCount를 long으로 변경
- DB count 결과를 int로 변환하지 않고 그대로 반환
- totalCount가 int 범위를 초과할 때 발생할 수 있는 변환 예외 제거

+++ aller-meal-api/src/main/java/com/allermeal/api/admin/response/AdminExternalApiLogPageResponse.java
@@ -7,7 +7,7 @@ public record AdminExternalApiLogPageResponse(
 	List<AdminExternalApiLogItemResponse> items,
 	int page,
 	int pageSize,
-	int totalCount
+	long totalCount
 ) {
 
 	public static AdminExternalApiLogPageResponse from(AdminExternalApiLogPageResult result) {
